Handle failed signature checks

+++ mousetracks2/utils/update.py
@@ -10,6 +10,8 @@ from urllib.request import urlopen
 from urllib.error import URLError
 
 from .network import safe_download_file
+from ..constants import UNTRUSTED_EXT
+from ..sign import verify_signature, failed_signature_verification
 from ..version import VERSION
 
 
@@ -90,27 +92,6 @@ def generate_exe_name(version: str = VERSION, with_extension: bool = True) -> st
     return f'MouseTracks-{version}-{_get_platform_suffix(with_extension)}'
 
 
-def _split_exe_name(path: str | Path) -> tuple[tuple[int, ...], str, str] | None:
-    """Get the components of the executable path.
-    Returns the version, os name and architecture.
-    """
-    if isinstance(path, Path) and not path.is_file():
-        return None
-
-    name = Path(path).stem
-    if not name.startswith('MouseTracks-'):
-        return None
-
-    parts = name.split('-')
-    try:
-        version = tuple(map(int, parts[1].split('.')))
-        os_name = parts[2]
-        arch = parts[3]
-    except (IndexError, ValueError):
-        return None
-    return version, os_name, arch
-
-
 def _get_platform_suffix(with_extension: bool = True) -> str:
     """Generate the file suffix from the platform data."""
     if sys.platform == 'win32':
@@ -137,24 +118,70 @@ def _get_platform_suffix(with_extension: bool = True) -> str:
     return f'{os_name}-{arch}{ext}'
 
 
-def download_version(folder: Path | str, version: str = 'latest') -> bool:
-    """Download a version of MouseTracks for the current OS to a file.
-    This will overwrite any existing files.
+def _split_exe_name(path: str | Path, include_untrusted: bool = False,
+                    ) -> tuple[tuple[int, ...], str, str] | None:
+    """Get the components of the executable path.
+    Returns the version, os name and architecture.
+    """
+    path = Path(path)
+    if not path.is_file():
+        return None
+
+    # Get any "unstrusted" files
+    if path.suffix == UNTRUSTED_EXT:
+        if not include_untrusted:
+            return None
+        path = path.with_suffix('')
+
+    name = path.stem
+    if not name.startswith('MouseTracks-'):
+        return None
+
+    parts = name.split('-')
+    try:
+        version = tuple(map(int, parts[1].split('.')))
+        os_name = parts[2]
+        arch = parts[3]
+    except (IndexError, ValueError):
+        return None
+    return version, os_name, arch
+
+
+def download_version(folder: Path | str, version: str = 'latest') -> Path | None:
+    """Download a version of MouseTracks for the current OS.
+    If the download fails the signature check, it will not be kept.
     """
     executable_link = get_download_link(version)
     if executable_link is None:
-        return False
+        return None
+
     executable_path = Path(folder) / Path(executable_link).name
+
+    # If it exists, then return without any extra checks
     if executable_path.exists():
-        return True
-    return safe_download_file(executable_link, executable_path)
+        return executable_path
+
+    # Check if signature verification previously failed to avoid redownloading
+    if failed_signature_verification(executable_path):
+        return None
+
+    if not safe_download_file(executable_link, executable_path):
+        return None
+
+    # Check the public key is a match
+    if verify_signature(executable_path):
+        return executable_path
+
+    # If verification fails, just remove the download
+    executable_path.unlink()
+    return None
 
 
 def cleanup_old_executables(folder: Path | str, version: str = VERSION, keep: int = 2) -> None:
     """Delete any executables older than the given version.
     Optionally define a number to keep.
     """
-    lower, current, higher = get_local_executables(folder, version)
+    lower, current, higher = get_local_executables(folder, version, include_untrusted=True)
 
     # Keep the highest few versions rather than removing everything
     if keep:
@@ -163,13 +190,21 @@ def cleanup_old_executables(folder: Path | str, version: str = VERSION, keep: in
     # Delete the files
     for executable in lower:
         print(f'Removing old executable: {executable}')
-        try:
-            executable.unlink()
-        except OSError as e:
-            print(f'Error removing file: {e}')
+        if executable.exists():
+            try:
+                executable.unlink()
+            except OSError as e:
+                print(f'Error removing file: {e}')
+
+        untrusted = executable.with_suffix(UNTRUSTED_EXT)
+        if untrusted.exists():
+            try:
+                untrusted.unlink()
+            except OSError as e:
+                print(f'Error removing file: {e}')
 
 
-def get_local_executables(folder: Path | str, version: str = VERSION,
+def get_local_executables(folder: Path | str, version: str = VERSION, include_untrusted: bool = False,
                           ) -> tuple[list[Path], Path | None, list[Path]]:
     """Get all the available executables from a path.
     Splits into lower and higher than the current version.
@@ -183,7 +218,7 @@ def get_local_executables(folder: Path | str, version: str = VERSION,
     executables_lower: list[tuple[tuple[tuple[int, ...], str, str], Path]] = []
     executables_higher: list[tuple[tuple[tuple[int, ...], str, str], Path]] = []
     for child in Path(folder).iterdir():
-        parts = _split_exe_name(child)
+        parts = _split_exe_name(child, include_untrusted)
         if parts is not None and parts[1] == result[1] and parts[2] == result[2]:
             if parts[0] < result[0]:
                 executables_lower.append((parts, child))
